Jenkins integration with cowbuilder
these two go together with a mvndput installation; these scripts
need to be placed in /usr/local/bin (currently hardcoded), and
the user Jenkins runs under needs to have password-less sudo on
kuhbauer. The /etc/pbuilderrc from MirDebian[1] is an absolute
requirement, and the base.cow-* must be set up by a sysadmin by
hand before and occasionally updated. Set DEBEMAIL in Jenkins,
an export statement in /etc/default/hudson seems to be a good
place. (Disable ENABLE_UPLOAD to work without mvndput.)

Currently, only Debian and *buntu suites are supported, not
Debian-Ports (which probably are broken in that pbuilderrc
anyway). Lists of suites, etc. are replicated several times
so take care. This is experimental, WFM quality.

[1] https://www.mirbsd.org/cvs.cgi/contrib/hosted/tg/deb/pbuilderrc?rev=HEAD